refactor(auth): extract AuthContext into separate file
