+++ frontend/src/contexts/auth.context.ts
@@ -1,0 +1,5 @@
+import { createContext } from 'react';
+import { AuthContextType } from '@/types/auth';
+
+// Create context with undefined initial value
+export const AuthContext = createContext<AuthContextType | undefined>(undefined);
