+++ frontend/src/hooks/useAuthContext.ts
@@ -1,5 +1,5 @@
 import { useContext } from 'react';
-import { AuthContext } from '@/contexts/AuthProvider';
+import { AuthContext } from '@/contexts/auth.context';
 
 export function useAuth() {
   const context = useContext(AuthContext);
